Scores calculated for two criteria
Updated map.java to have a population of road networks, to display the
best of them for visual purposes. Also calculates and keeps scores to
decide which one is 'the best'. Plus a bit of extra documentation in the
code.

TODO: the actual genetic algorithm - combining based on discussed ideas.
Potential issue: likely to often generate non-complete graphs despite
scoring system benefiting from complete graphs, due to somewhat random
nature of GA. This may be expected behaviour and may be the conclusion
of the paper. I'll try to finish a version of GA on Tuesday, but please
try on your own to see if you achieve better results.