...
Intersection maker made. But there's a serious bug in it somewhere...
and if you thought my code was ugly before, OO HOO, you'll have fun with
this.

I think a "mutator" needs to be added, otherwise I consider the code
done and ready for testing and writing a paper. If you want to do more,
that's up to you.