DONE!
Refactoring complete.
... as much as I'll do. It's 8am and I haven't slept in 24 hours.
Most code divided into classes/functions such that each function is
roughly 40 lines or less. Variable and function names taken into
consideration. At the very least, documentation at the beginning of each
function tells you all you need to know for the sake of writing for the
paper.

I already rewrote a bit of the old paper based on Goodwin's suggestions,
and will add a description of the genetic algorithm in next 18 hours.
Interesting part is to start collecting data to show, and analyzing it
properly. First, I need a nap...